Return the actual data type of the mutated property .. loaded properties are always represented as a float

+++ proc/src/test/java/org/neo4j/graphalgo/catalog/GraphSchemaWithMutationTest.java
@@ -67,7 +67,7 @@ public class GraphSchemaWithMutationTest extends BaseProcTest {
         assertCypherResult("CALL gds.graph.list() YIELD schema", singletonList(
             map(
                 "schema", map(
-                    "nodes", map("A", map("foo", "Float", "baz", "Float")),
+                    "nodes", map("A", map("foo", "Float", "baz", "Integer")),
                     "relationships", map("REL", map("bar", "Float"))
                 )
             )
